feat: add_robustness_tests
